+++ itour-parent/itour-model/src/main/java/com/itour/model/work/Like.java
@@ -22,13 +22,13 @@ public class Like extends Model<Like> {
      * 编号
      */
     @TableId(value = "ID", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 旅行信息编号
      */
     @TableField("WID")
-    private Integer wid;
+    private Long wid;
 
     /**
      * 用户编号
@@ -48,19 +48,19 @@ public class Like extends Model<Like> {
     @TableField("STATUS")
     private String status;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getWid() {
+    public Long getWid() {
         return wid;
     }
 
-    public void setWid(Integer wid) {
+    public void setWid(Long wid) {
         this.wid = wid;
     }
